byolinter uses only needed portions of PATH
